Moved logic from page into a child component
