+++ client/src/App.tsx
@@ -6,12 +6,13 @@ import HomePage from "./components/HomePage";
 import Layout from "./components/Layout";
 import QuizOverviewPage from "./components/QuizOverviewPage";
 import QuizPage from "./pages/QuizPage";
-import QuizLoginPage from "./pages/QuizLoginPage";
+import SlugValidation from "./components/Plastic/SlugValidation.tsx";
 import NotFound from "./pages/NotFound";
 import NavigationContext from "./context/NavigationContext";
 import PlasticProblemPage from "./pages/PlasticProblemPage.tsx";
 import PlasticConsequencesPage from "./pages/PlasticConsequencesPage.tsx";
 import PlasticApproachPage from "./pages/PlasticApproachPage.tsx";
+import QuizLoginPage from "./pages/QuizLoginPage.tsx";
 
 const App = () => {
   const [active, setActive] = useState<number>(0);
